requirements.txt
fastapi
uvicorn[standard]
langchain_huggingface
langchain_community
faiss-cpu
torch
transformers
google-generativeai
pydantic